requirements.txt
PyJWT~=2.8.0
bcrypt~=4.2.0
Flask~=3.0.3